*** empty log message ***
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@256 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/pss_full/pss_pss.c
@@ -833,7 +833,7 @@ int         sizename=0;
 int         err=0;
 char        test_name[200];
 int         dimensions[3];
-int         handle_dummy;
+long int    handle_dummy;
 #ifdef DEBUG 
 dstrc_enter("pss_chck");
 #endif
